+++ CHANGELOG.md
@@ -7,6 +7,10 @@ New features:
 - Added support for GPT-5.  Removed support for the deprecated GPT-4.5.
 - Added support for gpt-oss:20b and gpt-oss:120b via the Ollama driver.
 
+Bug fixes:
+
+- The filesystem tool was setting `0o600` permissions instead of `0o666 & ~umask`.
+
 ## v0.21 (2025-08-06)
 
 New features: